Remove wasm-bindings feature
