+++ Cargo.toml
@@ -36,15 +36,6 @@ std = [
     "url",
     "wasmtime",
 ]
-wasm-bindings = [
-    "futures-timer/wasm-bindgen",
-    "js-sys",
-    "rand/wasm-bindgen",
-    "send_wrapper",
-    "std",
-    "wasm-bindgen",
-    "web-sys",
-]
 
 [[bin]]
 name = "full-node"
@@ -111,12 +102,6 @@ soketto = { version = "0.4.2", optional = true }
 structopt = { version = "0.3.20", optional = true, default-features = false, features = ["color", "suggestions", "wrap_help"] }
 terminal_size = { version = "0.1.12", optional = true }
 url = { version = "2.1.1", optional = true }
-
-# `wasm-bindings` feature
-js-sys = { version = "0.3.45", optional = true }
-send_wrapper = { version = "0.5.0", optional = true }
-wasm-bindgen = { version = "0.2.68", optional = true }
-web-sys = { version = "0.3.45", optional = true, features = ["BinaryType", "CloseEvent", "DomException", "Event", "Storage", "WebSocket", "Window"] }
 
 # BELOW: DEPENDENCIES TO REMOVE
 # TODO:
