+++ src/database.rs
@@ -20,5 +20,5 @@
 //! This module contains sub-modules that provide different means of storing data in a
 //! persistent way.
 
+pub mod finalized_serialize;
 pub mod full_sled;
-pub mod local_storage_light;
